+++ engine/src/main/kotlin/engine/expressions/StatementExpressions.kt
@@ -3,7 +3,7 @@ package engine.expressions
 import engine.operators.Comparator
 import engine.operators.ComparisonOperator
 import engine.operators.DoubleComparisonOperator
-import engine.operators.EquationSystemOperator
+import engine.operators.StatementSystemOperator
 import engine.operators.StatementUnionOperator
 import engine.operators.StatementWithConstraintOperator
 import engine.sign.Sign
@@ -96,13 +96,13 @@ class StatementUnion(statements: List<Expression>, meta: NodeMeta = BasicMeta())
     val statements get() = children
 }
 
-class EquationSystem(equations: List<Expression>, meta: NodeMeta = BasicMeta()) : Expression(
-    operator = EquationSystemOperator,
+class StatementSystem(equations: List<Expression>, meta: NodeMeta = BasicMeta()) : Expression(
+    operator = StatementSystemOperator,
     operands = equations,
     meta = meta,
 ) {
     val equations get() = children
 
     fun withNamedEquations(name: (Int) -> String) =
-        equationSystemOf(equations.mapIndexed { i, eq -> eq.withName(name(i)) }).withOrigin(origin)
+        statementSystemOf(equations.mapIndexed { i, eq -> eq.withName(name(i)) }).withOrigin(origin)
 }