+++ engine/src/main/kotlin/engine/patterns/EquationPattern.kt
@@ -3,8 +3,7 @@ package engine.patterns
 import engine.context.Context
 import engine.expressions.Equation
 import engine.expressions.Expression
-import engine.operators.EquationSystemOperator
-import engine.operators.InequalitySystemOperator
+import engine.operators.StatementSystemOperator
 import engine.operators.StatementUnionOperator
 import engine.operators.StatementWithConstraintOperator
 
@@ -32,11 +31,9 @@ data class EquationPattern(
 
 fun commutativeEquationOf(lhs: Pattern, rhs: Pattern) = EquationPattern(lhs, rhs)
 
-fun equationSystemOf(eq1: Pattern, eq2: Pattern) = OperatorPattern(EquationSystemOperator, listOf(eq1, eq2))
+fun statementSystemOf(eq1: Pattern, eq2: Pattern) = OperatorPattern(StatementSystemOperator, listOf(eq1, eq2))
 
-fun inequalitySystemOf(eq1: Pattern, eq2: Pattern) = OperatorPattern(InequalitySystemOperator, listOf(eq1, eq2))
-
-fun equationUnionOf(eq1: Pattern, eq2: Pattern) = OperatorPattern(StatementUnionOperator, listOf(eq1, eq2))
+fun statementUnionOf(eq1: Pattern, eq2: Pattern) = OperatorPattern(StatementUnionOperator, listOf(eq1, eq2))
 
 fun statementWithConstraintOf(statement: Pattern, constraint: Pattern) =
     OperatorPattern(StatementWithConstraintOperator, listOf(statement, constraint))